converting favorite higher order component to hooks
